Color alpha changed to 0-1 range

+++ src/main/java/edu/uncc/cs/bridges_v21/base/ADTVisualizer.java
@@ -355,7 +355,7 @@ System.out.println(t_str);
 								Integer.toString(lv.getColor().getRed()) + COMMA +
 								Integer.toString(lv.getColor().getGreen()) + COMMA+
 								Integer.toString(lv.getColor().getBlue()) + COMMA +
-								Integer.toString(lv.getColor().getAlpha()) +
+								Float.toString(lv.getColor().getAlpha()) +
 							CLOSE_BOX + COMMA +
 							QUOTE + "thickness" + QUOTE + COLON + 
 							String.valueOf(lv.getThickness()) +
